Category is done(one-to-many relationship), database - fixed, minor bugs also fixed

+++ pet_shop/src/main/java/com/example/s13firstspring/models/entities/Category.java
@@ -17,9 +17,11 @@ import java.util.Set;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
     @Column
     private String name;
     @OneToMany(mappedBy="category")
     private Set<Subcategory> subcategories;
+
+
 }
